Make auth cookie name consistent.
We previously changed the name of the cookie that we expected to find, but not the name of the cookie that we set on login or cleared on logout! To avoid these kinds of problems in the future, use a static string for the cookie name and reference it rather than hard-coding it in the functions that interact with cookies.