Simplify top-level `Page` generation.
Rather than returning `maud::Markup` from top-level routes, return a
`bootstrap::Page`. This simplifies some of the logic in some of the
routes by interacting directly with `bootstrap::Page` (and, for example,
not having to pass `None` to an `Option<FlashMessage>` parameter). It
may also be more palatable to `rustfmt`.